This checkin represents CSL in a state of transition. The arrangements as regards FASL and !*SAVEDEF have been rationalised and the new scheme for creating checkpoint images is now the one in use. The garbage collector has been simplified so that the mark/slide version is no longer present - only the copying one is there. This last change means that the limit on the calculations that can be completed on a 32-bit machine will have been reduced, but I do not believe that this will be a serious inconvenience to anybody and the fact that the code is now shorter and hence easier to keep reliable may compensate.
The next steps will be (a) make FASL use the new serialization mechanism so
that there is just one body of  code for preserve/restart and FASL, (b) use
zlib to make image and FASL files smaller (I hope they will end up much smaller
and if I am lucky the saving will be almost a factor of two), (c) replace
parts of the garbage collector so that they do not need a mark bit, but instead
exloit the tag code for a forwarding pointer (that saves a bit in every
LispObject and removes some delicate assumptions made by the current code).

For the benefit of jlisp I have kept a copt of the previous version of
the Lisp compiler (which interfaces to the FASL scheme in a different way.
If I get keen or motivated and have not got other things to do I should
update Jlisp to use the newer mechanisms and then that can go away.