This represents quite a lot of re-work of the configuration and build scripts. That means that there is a risk that something will have ended up broken in the process - so PLEASE report to me if things go wrong.
Several things are going on in this change. Firstly the shell scripting in the
configure script now uses a few shell procedures - I hope that makes the
structure cleaner and the logic easier to follow. Next in various places I
try rather harder to get some of the autoreconf, configure and make steps
to exploit multi-core hosts better by running bits of themselves in
parallel. Especially on Windows this may speed builds up usefully. Then
on Windows (especially given that I am now aware that people still need 32-bit
platform support!) I try harder to build both 32 bit and 64-bit versions for
both native windows and cygwin. And I have TRIED To arrange things so that
any ill-advised person who confu¥igures (say) using a 32-bit cygwin shell and
later on goes "make" from a 64-bit cygwin shell will survive the experience.
I would tend to advise against that, obviously, but in the previous version
it happened accidentally the result could be a pretty bad mess.

For all of the Windows build to work to the FULL extent you will need 32-bit
cygwin installed in c:\cygwin and the 64-bit one in c:\cygwin64 (the default
locations for them).  You will want all development libraries etc installed
in both of those (see scripts/cygwin-sanity-check.sh for how to achieve that).
And all attempts to configure and build need to be from a cygwin shell (not,
for instance, and mingw one, even though that runs bash).

To survive fully fresh builds after "rm -rf cslbuild" I have provided a refresh
of csl/generated-c (the previous one was far enough out of step with current
CSL as to cause pain).

As an incidental change, there is a new cslbase/isprime.cpp that reduces the
size of tables that are used for checking numbers up to 2^64 to see if
they are prime.