Building Emacs REDUCE
In "esl.el":

Quote the list representation of a big integer used by the Calc library when returned by COMPRESS.  But this doesn't seem to work.  The fact that this representation of big integers does not evaluate to itself, as a number should, is a serious problem with using the Calc library and causes "sfpsi.red" to fail.  (It includes big integers in the source code, which I haven't noticed anywhere else.)

Modify GLOBAL to be callable at run time, as I modified FLUID some time ago.  This is necessary to build the tps package.

Allow CLOSE to quietly over-write a file in batch mode, whereas interactively it asks for confirmation.

Use the variable `noninteractive' directly in a couple of places rather than the more convoluted but consequential `(eq standard-output t)'.

Make LOAD-MODULE operate quietly under the control of a global variable.  This is essentially what the VERBOSELOAD switch does in PSL and perhaps I should use that mechanism instead.

In "eslrend.red":

Declare !*redefmsg to be fluid, which is necessary to build the qsum package.  REDEFMSG is a switch defined in PSL and used in "qsum.red".  In procedure `initreduce', do `!*int := not noninteractive', whereas previously I just set it to t.

In "reduce.el":

Load the core modules silently.  I still get a couple of warnings that I have not yet managed to suppress, but otherwise Emacs REDUCE now starts fairly tidily.

In "remake-one-package.sh":

Increase the maximum variable binding depth from the default of 1300 to 2000 in order to build the conlaw package.  (This problem with conlaw is not specific to Emacs Lisp!)


I now appear to be able to build all REDUCE 3.8 packages except defint, fide and fps, although sfpsi doesn't work and others also may not work.  defint and fps are broken as a consequence of big integers breaking sfpsi.  fide seems to be throwing syntax errors -- invalid comma -- probably caused by a declaration not working correctly.