Emacs REDUCE big integer support and build issues
In "esl.el":

Instead of quoting a big integer list representation when a big integer is generated by COMPRESS, make the representation self-quoting by defining bigpos and bigneg to be macros, in the same way that lambda is self-quoting in Emacs Lisp.

Turn off the irritating "working..." message generated by the Calc library.

Correct EXPT to generate a big integer instead of overflowing since a small integer power of a small integer can be quite big; e.g. in Emacs Lisp, 100^10 overflows and gives a negative result.

Fix EXPLODE2 to handle big integers.  This is used in the REDUCE output code.

Make *writingfaslfile upper case, since it is supposed to be the same variable as used in some REDUCE code where it gets converted to upper case.

Move the advice that makes EXPLODE use Emacs Lisp syntax from FASLOUT/FASLEND into the advice that makes the Lisp prettyprint function output to a fasl (source) file.  Also, stop the prettyprinter from outputting `(FASLEND)' when fasl input is interactive.  (There might be a more elegant way to do the latter, but if so I haven't worked out what it is yet.)

With the above changes, I can build and load the sfpsi module, and therefore also build the defint and fps modules, which depend on sfpsi.

In "eslrend.red":

Declare `noninteractive' to be global to avoid a warning message from REDUCE.  In fact, since this is an Emacs Lisp variable, I use the broken bar escape to avoid escaping every letter.  Maybe I should move this declaration to "esl.el".

In "reduce.el":

Add (put 'ARBINT 'SIMPFN 'SIMPIDEN) before loading the ALG package to avoid a warning message.  I think I may be doing things in a non-standard order, but as far as I can see ARBINT is only declared in ENTRY, which I load last of the core REDUCE packages.  Turn MSG off after loading ALG, which turns it on, and then turn MSG on after loading all the core packages, to avoid a warning about the FASLOUT stub not being defined in ENTRY.  FASLOUT and FASLEND are currently defined in "esl.el", but I might move them into a separate file, which I could then autoload more conventionally via ENTRY.  With these various hacks, REDUCE now loads cleanly with no spurious messages.

In "remake-one-package.sh":

Use multiple !s instead of the broken bar escape to avoid coding-system issues in a large number of log files.