Emacs REDUCE
Begin to use lexical binding instead of the default dynamic binding.  This is necessary to avoid some user variables getting mangled by being bound incorrectly.  It won't work (easily) for all of REDUCE but I'm hoping it will work for enough key components.  Until now, the variables u, v and x could not be used safely as symbolic-mode REDUCE variables because they picked up local bindings.  Using lexical binding for the rlisp files superv and rsupport (the latter only really for errorset*) seems to fix that problem.  It also means that the cali test now runs correctly.  I have yet to work out how to automate the use of lexical binding for appropriate files.

(Incidentally, the cali test runs about 100 times slower than it does in PSL on my computer, and I think Emacs REDUCE is consistently about 100 times slower than PSL REDUCE at present.  I'm hoping there is some scope for optimization!)

Replace the use of the non-ASCII broken bar character as a special escape and use the ASCII escape character instead.  Remove all other non-ASCII characters, none of which are really necessary.

Revert to proper definitions of RPLACA and RPLACD that don't do any copying.  (I changed them a while ago because I thought they might be causing problems.)

Extend the use of `cl-remprop' to remove properties and flags rather than setting the property value to nil.

Remove some obsolete commented-out code.

Define SET and RETURN as macros rather than aliases.  This change didn't solve any problems, but it seems to work and is consistent with the fact that SETQ and PROG seem to need to be macros.

Remove redundant function quotes ('#) in front of lambda expressions.  (I believe they are redundant in Emacs Lisp, although perhaps not in all Lisp dialects.  I was inconsistent about this.)

Replace a lambda application used as a let by a let, mainly to keep the compiler happy with lexical binding.  (Of course, REDUCE itself uses lambda like this a great deal!)

Tweak the print code to try harder to avoid outputting a space at the end of a line.  (I may need to revise this again later to bring it in line with the current behaviour of PSL and CSL.)

Put the read call back inside the ERRORSET in the main Standard Lisp read-eval-print loop.  (I think I took it out to facilitate debugging some time ago.)

Comment out the obsolete function MKFASL.  (I think this was used to build REDUCE 3.3.)

Remove the extra blank line that had started to appear between forms in the source files used to generate fasl files.  (I think this was a consequence of revising the print functions earlier.)