REDUCE on Common Lisp
Merge the case-inversion code into the Standard Lisp package and stop
using the CS-COMMON-LISP package and the CLISP `modern' facility.  Now
both SBCL and CLISP run essentially the same code with no discernible
time penalty due to case inversion.  This version runs more or less
equally correctly on both SBCL and CLISP, although in runs
significantly faster on SBCL.  It runs the REDUCE test files with only
a handful of glitches in some of the more obscure packages.

Handling of the escape character (!) is improved so that it appears in
output where it should.  Echo output is tracked so that the posn
function works more reliably.  I support *gc-hook* on SBCL (which
appears to be impossible on CLISP), so the timeouts in rubi_red work
on SBCL.