If you configure and build Reduce on a 64-bit windows machine running from the 64-bit version of cygwin then unless you specify --with-win32bit you should now only get 64-bit binaries created. There will still be several:   (1) a native windows one suitable for double clicking on that does NOT       create or remain attached to a console;   (2) A native windows version that is linked as a console mode application;   (3) A version that runs under cygwin64, either as a console application       or using X11 for its GUI.
Then on a different tack, I now make (almost) all items that will end up in
the heap get updated and accessed there using std::atomic<LispObject>*
rather than LispObject*. The purpose of this fairly dramatic change is so
that when and if I get a multi-threaded system working that all threads -
particularly ones that happen to be being executed on distinct CPU core4s -
all see a consistent view of the world and updates made by one will propagate
to the rest in a way that matches naive expectation. This does not put in
any critical regions to protect complicated shared structures. It mainly
addresses issues that arise with computers that can have hardware that
re-orders memory transactions in the name of performance. Under gcc on
cygwin what I see is that memory READs happen apparently using just the
same code that they used to. WRITES get a memory fence instruction adjacent
to them, or under clang++ on the Mac they may be implemeneted using an
atomic test-and-set instruction (but ignoring the test or exchange part of
that). On 32-bit ARM it looks as if there are more fences inserted, with
possibly some for memory reads. These changes will tend to hurt performance
a little, but on the PC the fact that read operations seem not to be changed
at all will surely mean that the incremental cost is pretty low.

There are some places where the code is conspicuously bad. For instance a
Lisp string in the heap is in effect an array of std::atomic<char>. If
I wanted to be really careful I would copy it out character by character
to a separate buffer to get a C-style string before using it with low-level
system facilities - but I tend (for instance when it is to be used as a
file-name?) to cast to a simple char * or const char *. Because std::atomic<T>
has standard layout etc I really rather expect this to be reasonable, but at
the very very least it violated C++ strict aliasing rules, and so the behaviour
of my code is officially undefined and in pratise it risks clever compilers
doing bad things with it. Ideally I should clean this up later on.

I DO NOT have multiple threads working and am quite some way from it, but
it I had not been able to make this change (and in fact it probably happened
more easily than I had expected!) then a pair of RPLACx operations done in one
thread might have ended up apparently performed in the other order when
viewed from a second thread (as memory writes propagated through cache
hardware in strage orders).