A general tidy up of the numerical code in sfellip.red and sfellipi.red in the specfn package.
Substantial changes to the theta functions code which now differ from those
originally defined by Lisa Temme in that the periods are 2*pi, 2*pi, pi and pi
respectively (NOT 4K, 4K, 2K and 2K).
Also the second argument is now the nome q where |q| < 1 and hence
the quasi-period is -i*log q (NB for q real log q < 0).
Again we are following the conventions of classical texts and, in particular,
DF Lawden: Elliptic Functions & Applications, Springer, 1989.

Implementation of generalised shift rules to capture the periodicity and
quasi-periodicity of the theta functions and so allowing all their zeros to be
recognised.

Use of vartheta rather than theta as the fancy-function symbol.

Substantial revisions to the numerical code for the theta functions including

a) corrections to the loop termination conditions to prevent premature
   loop exits when summing the defining series.

b) use of the Jacobi identity to speed up convergence when |q| is near to 1.
   However some issues regarding the multi-valued nature of theta1 and theta2
   remain with this when the nome is complex.

c) Some issues remain with rounding errors/slow-convergence when the first
   argument is complex with a large imaginary part owing to presence of
   hyperbolic functions. One should really use the quasi-periodicity to
   reduce the argument to lie in the region bounded by the rectangle
   (0, 0), (pi/2, 0), (pi/2, -i(log q/2)) and (0, -i(log q/2)).

Nevertheless the algorithm seems robust for real arguments and nome.

Some utility functions for evaluating the elliptic modulus and periods
of sn, cn and friends given the nome and conversely calculating the nome given
the modulus.

Additions to sfellip.tst and sfellip.rlg to include tests of these additions.