REDUCE web site
Use a more elegant font for REDUCE in the header and collapse the menu
into an icon on very small displays (i.e phones).  This is probably
not yet optimal but I think it's an improvement and I need to test it
live.

I now need to load the jQuery and bootstrap JS libraries on every
page, but I no longer load the Popper library, which I have never
actually used!  Hence, I no longer need the bootstrap-js.tpl file that
I introduced recently.

Revert the menu entry for the about.php page to "About REDUCE" because
"History and Contributors" overflows on narrow displays.