reduce-mode.el comment statement highlighting
I think I have finally fixed the font-lock highlighting of comment
statements by introducing a function to match them.  This needs to
look backwards to confirm that it has found a comment statement and I
can't see any way to do that using a single regexp.  Also,
consistently allow a comment statement to be terminated by either ; or
$.

Increase the package version to 1.55 with a view to a package release
in the not too distant future.  Remove tabs and tidy up some of the
formatting.