Overload mod to apply to both polynomials and integers
The infix operator mod has for some time had two conflicting
definitions, one in "poly/modular.red" and one in "poly/polydiv.red".
Now, mod is defined in "polydiv.red" (with higher precedence than in
"modular.red", immediately below plus) to call poly!-divide!*, which
calls evalmod if its arguments are both integers.  I have also
reinstated the Euclidean quotient operator that I originally called
div, but that clashes with other packages.  So I now call it
poly_quotient, which is not a great name but it is consistent with
pseudo_quotient.