This update to scripts/testall.sh and test1.sh alter the way to get testing of CSL varients other than the default one. They provide target options not just --csl but --csl-debug, --csl-nothrow, --csl-arithlib-debug and so on, basically all the suffixes that can appear in the cslbuild directory. So one can now for instance go   scripts/testall.sh --csl --csl-nothrow --psl --cslboot-debug to test all four versions - I might do that so that I could check correctness of the -nothrow results and observe the performance penalty that the cslboot-debug version suffers from. The total cululative time for all tests is stored in a "total.time" file in csl-times (etc).
While loking at this I observe that the sstools test script runs for
significantly longer than other tests and so anybody looking at the grand
total times will mainly be seeing the impact of that - which hardly seems fair.
I may in a later adjustment make the grand total exclude the time for that test
to get a more balances report?