oper.tex: Correct a typo. specfn.tex:  Minor corrections and updates to the Quick Reference tables.
alg/simp.red: add an extra test in simpiden and a support function
to check functions in the specfn and specfn2 packages are called
with the correct number of arguments. This avoids misleading output
and/or crashes in some fancy!-prifns if too few arguments are supplied.

alg/elem.red:  set number!-of!-args property to 1 for csc, sec, cot
and their hyperbolic analogues for more complete error checking.

misc/trigd/trigd.red:  Similar changes for secd, cscd and cotd.

Various files in specfn package: flag all user-level functions
in the specfn and specfn2 package as 'specfn and set the
number!-of!-args property appropriately to support changes in simpiden.

sfother.red: remove the rule for the integral of the Lambert_W function
as it is redundant.

sfairy.red & sfint.red: add prifns for Airy functions and integral functions
on non-graphics interfaces.