Assist section of HTML manual
Most of the text (from just below 16.5.3 Control of Switches) is
centred and clearly should not be. I think the cause was the use of
quotation environments, which I have replaced with quote environments
and this seems to fix the problem. I have also tweaked a few other
aspects of the formatting. The PDF version doesn't look much
different, just slightly more compact vertically, but the HTML version
looks much better.